Adds feature to allow for non-images to have hover listeners
That is, after using this library a couple of times, the most often
use-case I found was for buttons / links that contain an icon. It
was a little annoying that hovering over the entire link didn't cause
the image to change; I had to hover over the image itself, which if
its an icon, usually was a small part of the overall target.

This new feature changes that. When our event fires, we check to see
if the delegated target is an <img> tag. If not, we look for the first
child <img> tag to have its src swapped.

Although, it just now occurs to me, there may be cases where I don't
want the first img. This is definitely something I can optionally configure.